# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a assignment for DTS-Hacktiv8 final project.",
        "title": "Final Project API",
        "contact": {
            "name": "Aditya Gocendra",
            "email": "gocendra123@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/photos/": {
            "get": {
                "security": [
                    {
                        "Bearer \u003cJWT\u003e": []
                    }
                ],
                "description": "Note : To get photos, you need to send the token that has been obtained from the login response with the format Bearer \u003cYOUR TOKEN\u003e.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo"
                ],
                "summary": "Get Photo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetPhotoRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer \u003cJWT\u003e": []
                    }
                ],
                "description": "Note : To create photo, you need to send the token that has been obtained from the login response with the format Bearer \u003cYOUR TOKEN\u003e.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo"
                ],
                "summary": "Create Photo",
                "parameters": [
                    {
                        "description": "create user",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PhotoReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePhotoRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/photos/{photoid}": {
            "put": {
                "security": [
                    {
                        "Bearer \u003cJWT\u003e": []
                    }
                ],
                "description": "Note : To update photos, you need to send the token that has been obtained from the login response with the format Bearer \u003cYOUR TOKEN\u003e.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo"
                ],
                "summary": "Update Photo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "photoid",
                        "name": "photoid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update photo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PhotoReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePhotoRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer \u003cJWT\u003e": []
                    }
                ],
                "description": "Note : To delete photos, you need to send the token that has been obtained from the login response with the format Bearer \u003cYOUR TOKEN\u003e.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo"
                ],
                "summary": "Delete Photo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "photoid",
                        "name": "photoid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreatePhotoRes": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "photoUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ErrorMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "typeError": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetPhotoRes": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "photoUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/dto.PhotoUserRes"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "dto.PhotoReq": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.PhotoUserRes": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "dto.SuccessMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdatePhotoRes": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "photoUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer \u003cJWT\u003e": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}